Start Game Functionality & Basic Styling
Included buttons to start & reset the game + some basic styling choices
to display the board and player turn messages after the game has been
started. This code introduced a bug that broke the playRound Functionality
so the game no longer operates as expected after the board is shown. This
bug will be addressed in future commits.